Renamed rootValueStorage, moved into its own file. Avoid serializing an empty schema vector
